+++ waves/tests/test_builders.py
@@ -12,6 +12,12 @@ import SCons.Node.FS
 
 from waves import builders
 from waves._settings import _cd_action_prefix
+from waves._settings import _abaqus_environment_extension
+from waves._settings import _abaqus_datacheck_extensions
+from waves._settings import _abaqus_explicit_extensions
+from waves._settings import _abaqus_standard_extensions
+from waves._settings import _abaqus_solver_common_suffixes
+from waves._settings import _stdout_extension
 
 
 fs = SCons.Node.FS.FS()
@@ -47,6 +53,27 @@ def check_action_string(nodes, post_action, node_count, action_count, expected_s
         node.get_executor()
         assert len(node.executor.action_list) == action_count
         assert str(node.executor.action_list[0]) == expected_string
+
+
+def check_expected_targets(nodes, solver, stem):
+    """Verify the expected action string against a builder's target nodes
+
+    :param SCons.Node.NodeList nodes: Target node list returned by a builder
+    :param str solver: emit file extensions based on the value of this variable (standard/explicit/datacheck).
+    :param str stem: stem name of file
+
+    """
+    expected_suffixes = [_stdout_extension, _abaqus_environment_extension]
+    if solver == 'standard':
+        expected_suffixes.extend(_abaqus_standard_extensions)
+    elif solver == 'explicit':
+        expected_suffixes.extend(_abaqus_explicit_extensions)
+    elif solver == 'datacheck':
+        expected_suffixes.extend(_abaqus_datacheck_extensions)
+    else:
+        expected_suffixes.extend(_abaqus_solver_common_suffixes)
+    suffixes = [str(node).split(stem)[-1] for node in nodes]
+    assert set(expected_suffixes) == set(suffixes)
 
 
 prepend_env_input = {
@@ -281,19 +308,22 @@ def test_abaqus_solver_emitter(job_name, target, source, expected, outcome):
 # TODO: Figure out how to cleanly reset the construction environment between parameter sets instead of passing a new
 # target per set.
 abaqus_solver_input = {
-    "default behavior": ("abaqus", [], 7, 1, ["input1.inp"]),
-    "different command": ("dummy", [], 7, 1, ["input2.inp"]),
-    "post action": ("abaqus", ["post action"], 7, 1, ["input3.inp"])
+    "default behavior": ("abaqus", [], 7, 1, ["input1.inp"], None),
+    "different command": ("dummy", [], 7, 1, ["input2.inp"], None),
+    "post action": ("abaqus", ["post action"], 7, 1, ["input3.inp"], None),
+    "standard solver": ("abaqus", [], 8, 1, ["input4.inp"], "standard"),
+    "explicit solver": ("abaqus", [], 8, 1, ["input5.inp"], "explicit"),
+    "datacheck solver": ("abaqus", [], 11, 1, ["input6.inp"], "datacheck"),
 }
 
 
 @pytest.mark.unittest
-@pytest.mark.parametrize("abaqus_program, post_action, node_count, action_count, source_list",
+@pytest.mark.parametrize("abaqus_program, post_action, node_count, action_count, source_list, emitter",
                          abaqus_solver_input.values(),
                          ids=abaqus_solver_input.keys())
-def test_abaqus_solver(abaqus_program, post_action, node_count, action_count, source_list):
+def test_abaqus_solver(abaqus_program, post_action, node_count, action_count, source_list, emitter):
     env = SCons.Environment.Environment()
-    env.Append(BUILDERS={"AbaqusSolver": builders.abaqus_solver(abaqus_program, post_action)})
+    env.Append(BUILDERS={"AbaqusSolver": builders.abaqus_solver(abaqus_program, post_action, emitter)})
     nodes = env.AbaqusSolver(target=[], source=source_list, abaqus_options="")
     expected_string = f'cd ${{TARGET.dir.abspath}} && {abaqus_program} -information environment > ' \
                        '${job_name}.abaqus_v6.env\n' \
@@ -301,6 +331,7 @@ def test_abaqus_solver(abaqus_program, post_action, node_count, action_count, so
                        '${SOURCE.filebase} ${abaqus_options} -interactive -ask_delete no ' \
                        '> ${job_name}.stdout 2>&1'
     check_action_string(nodes, post_action, node_count, action_count, expected_string)
+    check_expected_targets(nodes, emitter, pathlib.Path(source_list[0]).stem)
 
 
 copy_substitute_input = {